refactor : NameCard 기타 파일들 네이밍 변경
