+++ feature/photocard/src/main/java/com/wepli/feature/photocard/detail/navigation/PhotoCardDetailNavigation.kt
@@ -5,7 +5,7 @@ import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
 import com.wepli.feature.photocard.detail.PhotoCardDetailScreenRoute
 import com.wepli.navigator.extras.Extras
-import com.wepli.navigator.feature.namecard.PhotoCardRoute
+import com.wepli.navigator.feature.photocard.PhotoCardRoute
 import com.wepli.shared.feature.uimodel.photocard.PhotoCardUiData
 import com.wepli.uimodel.music.SongUiData
 import extensions.enterAnimation
