+++ feature/photocard/src/main/java/com/wepli/feature/photocard/main/navigation/PhotoCardMainNavigation.kt
@@ -4,7 +4,7 @@ import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
 import com.wepli.feature.photocard.main.PhotoCardScreenRoute
-import com.wepli.navigator.feature.namecard.PhotoCardRoute
+import com.wepli.navigator.feature.photocard.PhotoCardRoute
 
 // Controller
 fun NavController.navigateToPhotoCardMain() {
